requirements.txt
Flask==3.0.3
flask-cors==4.0.0
pypdf==4.2.0
Flask-Session>=0.5.0
google-generativeai>=0.3.0
gunicorn==21.2.0

